Fixed issue with double announcement from Narrator
Issue is reproduced due to incorrect behavior of Narrator when we call "RaiseAutomationEvent(UIA.Text_TextChangedEventId);" method. As result we have an issue with blue rectangle and double announcement in Windows 1809.

Added calling of the "RaiseAutomationNotification" instead of "RaiseAutomationEvent" method. This method contains the beginning of the standard announcement from Narrator. Ending of announcement is called by assigning a value.

Also added condition for Focus method because calling this method initiates additional announcement from Narrator on click by mouse